feat: Added backend data, logging, and actionable execution
This commit introduces several key features and improvements:

-   **Backend Data:** Added sample JSON files
 (`backend_request.json`, `backend_response_data.json`, `backend_response_battery.json`, `backend_response_structure.json`) to simulate backend requests and responses.
-   **Logging Interceptor:** Implemented `LoggingInterceptor` to log both request and response bodies in a
 human-readable JSON format for debugging. The `LoggingInterceptor` is added to `OkHttpClient`
-   **Actionable Execution:** Added `ActionableExecutor` and integrated its execution in `DashBoardViewModel`.
-    Refactor: added Response: json() method explanation

+++ app/src/main/java/com/hackathon/powergaurd/data/remote/PowerGuardRepository.kt
@@ -164,17 +164,17 @@ class PowerGuardRepository @Inject constructor(
 
         return AnalysisResponse(
             id = deviceData.deviceId,
-            timestamp = System.currentTimeMillis(),
-            batteryScore = batteryScore,
-            dataScore = dataScore,
-            performanceScore = performanceScore,
+            timestamp = System.currentTimeMillis().toFloat(),
+            batteryScore = batteryScore.toFloat(),
+            dataScore = dataScore.toFloat(),
+            performanceScore = performanceScore.toFloat(),
             insights = insights,
             actionable = actionable,
             success = true,
             message = "Analysis",
             estimatedSavings = AnalysisResponse.EstimatedSavings(
-                batteryMinutes = 17,
-                dataMB = 1200
+                batteryMinutes = 17f,
+                dataMB = 1200f
             )
         )
     }
